requirements.txt
django
django-crispy-forms
django-picklefield
RestrictedPython